Update development scripts and dependencies
- Modified the `dev` script in `package.json` to target the `examples` filter for database pushing.
- Added new `db:push` and `db:studio` scripts in the `examples/package.json` for managing the database with `drizzle-kit`.
- Introduced `drizzle.config.ts` and `postgres-schema.ts` in the `examples` directory to define database schema and configuration.
- Updated `pnpm-lock.yaml` to include new dependencies for `drizzle-orm`, `postgres`, and `drizzle-kit`.
- Removed obsolete Postgres provider implementation and related files from the `packages/docsync` directory.

These changes streamline the development process and enhance database management capabilities.
