+++ .github/workflows/playwright.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Install dependencies
         run: npm install -g pnpm && pnpm install
       - name: Push database schema
-        run: pnpm --filter @docukit/docsync db:push
+        run: pnpm --filter examples db:push
       - name: Create examples .env
         run: echo "DOCNODE_DB_URL=$DOCNODE_DB_URL" > examples/.env
       - name: Build examples and workspace dependencies
