+++ examples/docsync-server.ts
@@ -1,7 +1,8 @@
 import { DocNodeBinding } from "@docukit/docsync-react/docnode";
-import { DocSyncServer, postgresProvider } from "@docukit/docsync-react/server";
+import { DocSyncServer } from "@docukit/docsync-react/server";
 import { indexDocConfig } from "./shared-config.ts";
 import { lexicalDocNodeConfig } from "@docukit/docnode-lexical";
+import { postgresProvider } from "./postgres-provider.ts";
 
 new DocSyncServer({
   docBinding: DocNodeBinding([indexDocConfig, lexicalDocNodeConfig]),
