+++ examples/postgres-schema.ts
@@ -1,0 +1,26 @@
+import {
+  jsonb,
+  pgTable,
+  primaryKey,
+  timestamp,
+  varchar,
+} from "drizzle-orm/pg-core";
+
+export const documents = pgTable("_docsync_documents", {
+  userId: varchar("userId", { length: 26 }).notNull(),
+  docId: varchar("docId", { length: 26 }).notNull().primaryKey(),
+  doc: jsonb("doc").notNull(),
+  clock: timestamp("clock", { precision: 3, withTimezone: true }).notNull(),
+});
+
+export const operations = pgTable(
+  "_docsync_operations",
+  {
+    docId: varchar("docId", { length: 26 }).notNull(),
+    operations: jsonb("operations").notNull(),
+    clock: timestamp("clock", { precision: 3, withTimezone: true })
+      .notNull()
+      .defaultNow(),
+  },
+  (table) => [primaryKey({ columns: [table.docId, table.clock] })],
+);
