+++ packages/docsync/src/exports/server.ts
@@ -7,5 +7,4 @@ export type {
   ClientDisconnectEvent,
   SyncRequestEvent,
 } from "../server/types.js";
-export { postgresProvider } from "../server/providers/postgres/index.js";
 export { inMemoryServerProvider } from "../server/providers/memory.js";
